lib: fix unlz4 pointer mismatch

+++ lib/decompress_unlz4.c
@@ -31,10 +31,10 @@
 #define LZ4_DEFAULT_UNCOMPRESSED_CHUNK_SIZE (8 << 20)
 #define ARCHIVE_MAGICNUMBER 0x184C2102
 
-STATIC inline int INIT unlz4(u8 *input, long in_len,
-				long (*fill)(void *, unsigned long),
-				long (*flush)(void *, unsigned long),
-				u8 *output, long *posp,
+STATIC inline int INIT unlz4(u8 *input, int in_len,
+				int (*fill) (void *, unsigned int),
+				int (*flush) (void *, unsigned int),
+				u8 *output, int *posp,
 				void (*error) (char *x))
 {
 	int ret = -1;
